third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-rename-mapped-file.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../debugger-test.js"></script>
<script src="../isolated-filesystem-test.js"></script>
<script src="./persistence-test.js"></script>
<script src="./resources/foo.js"></script>
<script>

function test() {
  BindingsTestRunner.forceUseDefaultMapping();
  var fs = new BindingsTestRunner.TestFileSystem('file:///var/www');
  var fsEntry = BindingsTestRunner.addFooJSFile(fs);
  fs.addFileMapping('http://127.0.0.1:8000', '/');
  fs.reportCreated(function() {});
  BindingsTestRunner.waitForBinding('foo.js').then(onBindingCreated);
  Persistence.persistence.addEventListener(Persistence.Persistence.Events.BindingRemoved, onBindingRemoved);

  function onBindingCreated(binding) {
    TestRunner.addResult('Binding created: ' + binding);
    TestRunner.addResult('Renaming foo.js -> bar.js');
    binding.fileSystem.rename('bar.js', function() {});
  }

  function onBindingRemoved(event) {
    var binding = event.data;
    TestRunner.addResult('Binding successfully removed: ' + binding);
    TestRunner.completeTest();
  }
};
</script>
</head>
<body onload="runTest()">
<p>Verify that binding gets removed as the fileSystem file gets renamed.</p>
</body>
</html>
