third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-rename-mapped-file-expected.txt

Verify that binding gets removed as the fileSystem file gets renamed.

Binding created: {
       network: http://127.0.0.1:8000/inspector/persistence/resources/foo.js
    fileSystem: file:///var/www/inspector/persistence/resources/foo.js
}
Renaming foo.js -> bar.js
Binding successfully removed: {
       network: http://127.0.0.1:8000/inspector/persistence/resources/foo.js
    fileSystem: file:///var/www/inspector/persistence/resources/bar.js
}
